% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ct.ps.r
\name{get.ct.ps}
\alias{get.ct.ps}
\title{Get celltype p-values}
\usage{
get.ct.ps(cellTypesToConsider, specificity, hitGenes, patchCellI, patchSeqExp,
  bg)
}
\arguments{
\item{cellTypesToConsider}{List of strings for cell type names from the reference dataset}

\item{specificity}{A specificity matrix, generally for level2 celltypes, obtained using EWCE (only the specificity matrix (not the ctd) should be passed)}

\item{hitGenes}{Genes which fulfull the following criteria: (1) Expressed in the reduced patchseq dataset, i.e. they have standard deviation over a given 
threshold; (2) They have at least one read in patchseq cell 'i'; (3) They are found in both patchseq and specificity datasets}

\item{patchCellI}{Index of which patchseq cell is being considered}

\item{patchSeqExp}{Gene by Cell matrix of the patchseq data}

\item{bg}{Vector of gene symbols. All of these must be in the specificity matrix.}
}
\value{
A vector of p-values corresponding to cellTypesToConsider, where lowest value indicates that celltype is the most likely type for cell i
}
\description{
Evaluates for cell type 'i' which of the reference cell types it is most likely to be. The probability values are returned as a vector.
}
\examples{
plot_crayons()

}
