% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.patchseq.r
\name{map.patchseq}
\alias{map.patchseq}
\title{Map PatchSeq}
\usage{
map.patchseq(ctd, patchdata, sdThresh = 1, annotLevel = 2,
  useSDthresh = TRUE, reps = 10000)
}
\arguments{
\item{ctd}{Cell type specificity data as generated by EWCE}

\item{patchdata}{Gene by cell matrix of patchseq data}

\item{sdThresh}{Genes from the patchseq data will only be kept if they have a standard devation across the dataset larger than this figure}

\item{annotLevel}{Annotation level of ctd to use}

\item{useSDthresh}{Should standard deviation or total reads be used to determine which patchseq genes to keep?}

\item{reps}{How many bootstrap replicates to use. Ideally 10000.}
}
\value{
A list with assigned and fullRes
}
\description{
Takes celltype specificity data ('ctd') and patchseq expression data ('patchdata') and returns which cell type each patchseq cell maps onto
}
\examples{
map.patchseq(ctd,patchdata_Reduced,sdThresh=1,annotLevel=2,useSDthresh=TRUE,reps=10000)

}
